Closes the 2026-06 workspace review's remaining P1s plus the pulled-in
Ziv-depth P2 family, in six slices on this branch. The arc's center of
gravity: input-encoded depth and rim/ceiling integer arithmetic — the
two ways a kernel's resolution model and the representation's edges
manufactured certified wrong answers.

R2.1 (ADR-0102): hypot/atan/atan2 dispatch input-encoded depth through
round_with_infinitesimal with a grid-derived boundary-gap argument
(atan was pf-e2ow's actual root: no tiny-x dispatch at all). Closed
pf-71u2, pf-e2ow.

R2.2 (ADR-0103): the Ziv driver takes a lazy input-derived
certification depth — one certified deep rung past the legacy cap on
exhaustion; zeta and the six trig kernels adopt it; the trig range cap
decoupled from working precision. Closed pf-jl35.

R2.2b (ADR-0104): the ADR-0059 tiny-x family gains input-precision
arms and depth hints (floors one series term deeper after the
terminating-correction refutation); the expm1/log1p internal boost
caps removed (the arm alone would have resurrected certified zeros);
hypot gains a construct-and-check exactness audit. Closed pf-fbjn
(found by R2.1's own derivation).

R2.3 (ADR-0105): acos gains the asin gap boost (named reproducer was
+0 with Status OK); the INEXACT force covers Zero results class-wide
through one helper, with the per-kernel exact-zero audit recorded;
li/Ei/Ci transcendence overclaims corrected to the conditional
posture. Closed pf-9761.

R2.4 (ADR-0106): agm normalizes toward exponent 0 below spread 2^33
and takes the certified asymptotic closed form above it (the iterates
converge to the AGM's exponent, so no static normalization survives a
~2^63 spread); the scale-back surfaces its saturation. Closed
pf-06lk.

R2.5 (ADR-0107/0108): round_with_infinitesimal lifts away from the
bottom rim (the ADR-0102 interim rim guards removed; the bottom-most
borrow gets the mode-aware underflow dispatch); the Taylor loops stop
adding rim-saturated garbage terms; the half-width hardened; the deep
rung gains a feasibility ceiling; parse and thirteen ops/rounding
sites get consistent u32-ceiling arithmetic (source-clamped widths,
u64 placement shapes). Closed pf-a77o, pf-9wb2.

Method per slice: mpmath-pinned red lane rows -> grid-checked
derivation -> independent adversarial verifier -> full CI-parity gate
battery -> unsigned commit. The verifiers refuted first drafts
nineteen times across the arc; every refutation was fixed pre-commit
and recorded in the owning ADR's inversion section. Found-and-filed
for later arcs: pf-hkoj (zeta_fe near trivial zeros, P1), pf-0r1l
(li/Ei/digamma deep interior zeros, P1), pf-31ql (Si tiny-x), pf-zmk3
(rim-saturated-zero flags).

Lane: 37 -> 66 rows (64 debug + 2 release-gated); gate battery 31/31
per slice; consumer suites (ball, complex, libm differential-mpfr)
green throughout.

+++ src/mantissa.rs
@@ -59,7 +59,11 @@ pub(crate) trait Mantissa {
 #[inline]
 #[must_use]
 pub(crate) const fn storage_shift(limbs: usize, value_bits: u32) -> u32 {
-    (limbs as u32) * 64 - value_bits
+    // u64 arithmetic (pf-9wb2, ADR-0108): `limbs · 64` exceeds u32
+    // for value_bits within 63 of the precision ceiling — a debug
+    // overflow panic at mere construction (release was
+    // self-correcting modular arithmetic). The difference is < 64.
+    ((limbs as u64) * 64 - value_bits as u64) as u32
 }
 
 #[cfg(test)]
